audit-compliance-app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pyyaml==6.0.2
reportlab==4.2.2
